% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betahyperMH.R
\name{betahyperMH}
\alias{betahyperMH}
\title{Metropolis-Hastings update for the beta prior hyper parameters}
\usage{
betahyperMH(a, b, atune, btune, phivec)
}
\arguments{
\item{a}{current a.phi parameter to be updated}

\item{b}{current b.phi parameter to be updated}

\item{atune}{Tuning parameter one}

\item{btune}{Tuning parameter two}

\item{phivec}{vector of current values of phi for each location}
}
\value{
a numeric vector
}
\description{
Metropolis-Hastings update for beta prior hyper-parameters
}
