% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_chain.R
\name{gibbs_chain}
\alias{gibbs_chain}
\title{Runs a gibbs sampler chain for capture recapture data}
\usage{
gibbs_chain(dataList, seed, N = 10000, atune = 0.25, btune = 1,
  parnames = NULL)
}
\arguments{
\item{dataList}{list containing a locList for each location}

\item{seed}{a seed}

\item{N}{number of iterations for the gibbs sampler, default is 10,000}

\item{atune}{Tuning parameter for the Metropolis-Hastings update of beta hyperparameters. Default is 0.25.}

\item{btune}{Tuning parameter for Metropolis-Hastings update of beta hyperparamters. Defaults is 1.}

\item{parnames}{Vector containing all the parameter names, must coincide with the parameters in dataList}
}
\value{
a numeric matrix
}
\description{
Runs a gibbs chain for capture recapture data
}
