% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_update.R
\name{loc_update}
\alias{loc_update}
\title{Workhouse function that updates previous iterations estimates for a single location}
\usage{
loc_update(locList, a.phi, b.phi)
}
\arguments{
\item{locList}{list containing a data frame and current parameters for each location}

\item{a.phi}{alpha hyperparameter for beta prior for overall phi}

\item{b.phi}{beta hyperparameter for beta prior for overall phi}
}
\value{
a numeric matrix
}
\description{
updates parameters for a single location
}
