% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_posterior.R
\name{make_posterior}
\alias{make_posterior}
\title{Create posterior distribution for parameters by combining all gibbs chains}
\usage{
make_posterior(chainList)
}
\arguments{
\item{chainList}{list where each element is an output chain from gibbs_chain()}
}
\description{
Create posterior distribution for parameters by combining all gibbs chains
}
