% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postlike.R
\name{postlike}
\alias{postlike}
\title{calculate the posterior likelihood for betahyperMH Metropolis update}
\usage{
postlike(x, y, phivec)
}
\arguments{
\item{x}{log(a/(a+b)) calculated in betahyperMH}

\item{y}{log(a+b) calculated in betahyperMH}

\item{phivec}{vector of current values of phi for each location}
}
\value{
a single numeric value
}
\description{
calculate the posterior likelihood for betahyperMH
}
